move around the gui package because it was ugly
