+++ src/main/java/io/github/codeutilities/commands/item/TemplatesCommand.java
@@ -6,7 +6,7 @@ import com.mojang.brigadier.CommandDispatcher;
 import io.github.codeutilities.CodeUtilities;
 import io.github.codeutilities.commands.Command;
 import io.github.codeutilities.commands.arguments.ArgBuilder;
-import io.github.codeutilities.gui.TemplateStorageUI;
+import io.github.codeutilities.gui.menus.TemplateStorageUI;
 import io.github.codeutilities.util.chat.ChatType;
 import io.github.codeutilities.util.chat.ChatUtil;
 import io.github.codeutilities.util.networking.WebUtil;
