+++ src/main/java/io/github/codeutilities/commands/util/PlotsCommand.java
@@ -3,7 +3,7 @@ package io.github.codeutilities.commands.util;
 import com.mojang.brigadier.CommandDispatcher;
 import io.github.codeutilities.commands.Command;
 import io.github.codeutilities.commands.arguments.ArgBuilder;
-import io.github.codeutilities.gui.PlotsStorageUI;
+import io.github.codeutilities.gui.menus.PlotsStorageUI;
 import io.github.codeutilities.util.ItemUtil;
 import io.github.codeutilities.util.file.ExternalFile;
 import io.github.cottonmc.clientcommands.CottonClientCommandSource;
