+++ src/main/java/io/github/codeutilities/gui/menus/ItemEditorGui.java
@@ -1,6 +1,8 @@
-package io.github.codeutilities.gui;
+package io.github.codeutilities.gui.menus;
 
 import io.github.codeutilities.CodeUtilities;
+import io.github.codeutilities.gui.widgets.CItem;
+import io.github.codeutilities.gui.widgets.CTextField;
 import io.github.codeutilities.util.IMenu;
 import io.github.codeutilities.util.StringUtil;
 import io.github.cottonmc.cotton.gui.client.LightweightGuiDescription;
