+++ src/main/java/io/github/codeutilities/gui/menus/TemplateSearchGui.java
@@ -1,8 +1,10 @@
-package io.github.codeutilities.gui;
+package io.github.codeutilities.gui.menus;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import io.github.codeutilities.gui.widgets.ClickableGiveItem;
+import io.github.codeutilities.gui.widgets.ItemScrollablePanel;
 import io.github.codeutilities.util.templates.TemplateUtils;
 import io.github.cottonmc.cotton.gui.client.LightweightGuiDescription;
 import io.github.cottonmc.cotton.gui.widget.WGridPanel;
