+++ src/main/java/io/github/codeutilities/gui/widgets/CItem.java
@@ -1,4 +1,4 @@
-package io.github.codeutilities.gui;
+package io.github.codeutilities.gui.widgets;
 
 import io.github.cottonmc.cotton.gui.widget.TooltipBuilder;
 import io.github.cottonmc.cotton.gui.widget.WItem;
