+++ src/main/java/io/github/codeutilities/gui/widgets/CTextField.java
@@ -1,4 +1,4 @@
-package io.github.codeutilities.gui;
+package io.github.codeutilities.gui.widgets;
 
 import io.github.cottonmc.cotton.gui.widget.WTextField;
 import net.minecraft.text.Text;
